requirements.txt
streamlit
requests
Pillow
pandas
fastapi[standard]
pinecone
transformers 